requirements.txt
langgraph>=0.2.0
langchain-core>=0.2.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0
langfuse>=2.0.0
reportlab>=4.0.0
python-abc
pymongo
mysql-connector-python
boto3
dotenv
